leak fixes

+++ src/execute_tree.c
@@ -6,7 +6,7 @@
 /*   By: fernacar <fernacar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/03 18:36:37 by fernacar          #+#    #+#             */
-/*   Updated: 2023/10/18 17:47:49 by fernacar         ###   ########.fr       */
+/*   Updated: 2023/10/23 19:53:36 by fernacar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,20 +17,23 @@ static char	*get_full_path(char *command)
 	char	*ret;
 	char	*temp;
 	char	**paths;
+	int		i;
 
 	ret = NULL;
 	paths = ft_split(getenv("PATH"), ':');
-	while (*paths != NULL)
+	i = 0;
+	while (paths[i] != NULL)
 	{
-		temp = ft_strjoin(*paths, "/");
+		temp = ft_strjoin(paths[i], "/");
 		ret = ft_strjoin(temp, command);
 		free(temp);
 		if (access(ret, X_OK) == 0)
  			break ;
 		free(ret);
 		ret = NULL;
-		paths++;
+		i++;
 	}
+	free_split(paths);
 	return (ret);
 }
 
@@ -44,7 +47,8 @@ static void	execute_command(char **args, char **envp)
 		if (access(full_path, X_OK) != 0)
 		{
 			perror(full_path);
-			exit(1);	// check if correct
+			free(full_path);
+			exit(127);
 		}
 	}
 	else
@@ -53,12 +57,20 @@ static void	execute_command(char **args, char **envp)
 		if (full_path == NULL)
 		{
 			dprintf(2, "%s: command not found\n", args[0]);
-			exit(1);	// check if correct
+			// free_split(args);
+			exit(127);
 		}
 	}
-	execve(full_path, args, envp);
-	dprintf(2, "exec %s failed :(\n", args[0]);
+	if (!fork1())
+	{
+		execve(full_path, args, envp);
+		free(full_path);
+		free_split(args);
+		panic("execve error");
+	}
+	wait(0);
 	free(full_path);
+	// free_split(args);
 }
 
 void	execute_node(t_tnode *node, char** envp)
@@ -74,17 +86,19 @@ void	execute_node(t_tnode *node, char** envp)
 		panic("null node");
 	if (node->type == EXEC)
 	{
-		if (!(pid = fork1()))
-		{
-			exec_node = (t_tnode_exec*)node;
+		// ft_putstr_fd("EXEC\n", 2);
+		exec_node = (t_tnode_exec*)node;
+		// if (!(pid = fork1()))
+		// {
 			if (exec_node->argv[0] == 0)
 				exit(1);
 			execute_command(exec_node->argv, envp);
-		}
-		waitpid(pid, 0, 0);
+		// }
+		// waitpid(pid, 0, 0);
 	}
 	else if (node->type == REDIR)
 	{
+		// ft_putstr_fd("REDIR\n", 2);
 		redir_node = (t_tnode_redir*)node;
 		close(redir_node->fd);
 		if (open(redir_node->file, redir_node->flags, 0666) < 0)
@@ -92,10 +106,12 @@ void	execute_node(t_tnode *node, char** envp)
 			perror(redir_node->file);
 			exit(1);
 		}
-		execute_node(redir_node->node, envp);
+		if (redir_node->node)
+			execute_node(redir_node->node, envp);
 	}
 	else if (node->type == PIPE)
 	{
+		// ft_putstr_fd("PIPE\n", 2);
 		pipe_node = (t_tnode_pipe*)node;
 		if (pipe(p) < 0)
 			panic("pipe error");
@@ -104,14 +120,16 @@ void	execute_node(t_tnode *node, char** envp)
 			dup2(p[1], STDOUT_FILENO);
 			close(p[1]);
 			close(p[0]);
-			execute_node(pipe_node->branch_left, envp);
+			if (pipe_node->branch_left)
+				execute_node(pipe_node->branch_left, envp);
 		}
 		if (fork1() == 0)
 		{
 			close(p[1]);
 			dup2(p[0], STDIN_FILENO);
 			close(p[0]);
-			execute_node(pipe_node->branch_right, envp);
+			if (pipe_node->branch_right)
+				execute_node(pipe_node->branch_right, envp);
 		}
 		close(p[0]);
 		close(p[1]);
@@ -120,30 +138,35 @@ void	execute_node(t_tnode *node, char** envp)
 	}
 	else if (node->type == HEREDOC)
 	{
+		// ft_putstr_fd("HEREDOC\n", 2);
 		heredoc_node = (t_tnode_heredoc*)node;
 		char *line;
 
-		int fd_temp = open(".heredoc", O_WRONLY | O_CREAT | O_TRUNC, 0600);
+		int fd = open(".heredoc", O_WRONLY | O_CREAT | O_TRUNC, 0600);
 		while (1)
 		{
 			line = readline("heredoc> ");
-			if (!line || ft_strcmp(line, heredoc_node->delm_start) == 0)
+			if (!line || ft_strcmp(line, heredoc_node->delm) == 0)
 			{
 				if (line)
 					free(line);
 				break ;
 			}
-			ft_putstr_fd(line, fd_temp);
-			ft_putstr_fd("\n", fd_temp);
+			ft_putstr_fd(line, fd);
+			ft_putstr_fd("\n", fd);
 		}
-		close(fd_temp);
+		//close(fd_temp);
 
-		close(STDIN_FILENO);
-		fd_temp = open(".heredoc", O_RDONLY, 0400);
-		execute_node(heredoc_node->node, envp);
+		// close(STDIN_FILENO);
+		// fd_temp = open(".heredoc", O_RDONLY, 0400);
+		if (dup2(fd, STDIN_FILENO) == -1)
+			panic("dup2 error");
+		close(fd);
+		if (((t_tnode_exec *)(heredoc_node->node))->argv[0] != 0)
+			execute_node(heredoc_node->node, envp);
 		unlink(".heredoc");
 	}
 	else
 		panic("invalid node type");
-	exit(0);
+	// exit(0);
 }
